Move JetStream tests to own directory
