+++ tests/Feature/JetStream/EmailVerificationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\JetStream;
 
 use App\Models\User;
 use App\Providers\RouteServiceProvider;
