+++ tests/Feature/JetStream/TwoFactorAuthenticationSettingsTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\JetStream;
 
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
